requirements.txt
langchain
langchain_core
langchain-openai
langchain-chroma
pandas
ipykernel
python-dotenv
fire